+++ display/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Add more getters for content, column, and double-column widths.
+- Allow overriding the material suggested margins and gutters.
+
 ## 0.1.0
 
 - Update with latest material design guidelines.
@@ -5,8 +10,8 @@
 
 ## [0.0.2] - August 25, 2018
 
-* Fix docs.
+- Fix docs.
 
 ## [0.0.1] - August 25, 2018
 
-* First release.
+- First release.